requirements.txt
pymupdf==1.24.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
streamlit==1.36.0
python-dotenv==1.0.1
reportlab==4.2.0
pytest==8.2.0